refactor: 공통과정, 학부과정, 대학원과정 상속 구조 도입
