+++ backend/src/main/java/org/example/backend/course/controller/CourseController.java
@@ -5,12 +5,18 @@ import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.course.domain.dto.CourseReqDto;
-import org.example.backend.course.domain.dto.CourseResDto;
+import org.example.backend.course.domain.dto.res.CourseResDto;
 import org.example.backend.course.domain.entity.CourseType;
 import org.example.backend.course.service.CourseService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
