+++ backend/src/main/java/org/example/backend/course/domain/entity/BaseCourse.java
@@ -1,0 +1,47 @@
+package org.example.backend.course.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.DiscriminatorColumn;
+import jakarta.persistence.DiscriminatorType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.InheritanceType;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.example.backend.common.domain.BaseEntity;
+
+@Entity
+@Getter
+@Inheritance(strategy = InheritanceType.JOINED)
+@DiscriminatorColumn(name = "course_type", discriminatorType = DiscriminatorType.STRING)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public abstract class BaseCourse extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "course_id", nullable = false)
+    private Long id;
+
+    @Column(name = "course_number", nullable = false)
+    protected String courseNumber;
+
+    @Column(name = "course_name", nullable = false)
+    protected String courseName;
+
+    @Column(name = "course_name_en")
+    protected String courseNameEn;
+
+    @Column(name = "credit_time")
+    protected String creditTime;
+
+    protected BaseCourse(String courseNumber, String courseName, String courseNameEn, String creditTime) {
+        this.courseNumber = courseNumber;
+        this.courseName = courseName;
+        this.courseNameEn = courseNameEn;
+        this.creditTime = creditTime;
+    }
+}
